requirements_clean.txt
# Core Django
Django==5.1.7
djangorestframework==3.15.2
dj-database-url==2.3.0

# Channels and WebSockets
channels==4.2.2
channels_redis==4.2.1
daphne==4.1.2

# Database
PyMySQL==1.1.1
psycopg2-binary==2.9.10

# Forms and UI
django-crispy-forms==2.3
crispy-bootstrap5==2024.10
django-widget-tweaks==1.5.0

# Media and Files
Pillow==11.1.0
whitenoise==6.9.0

# Email processing
imap-tools==1.7.2

# Scheduling
schedule==1.2.2

# Redis
redis==6.2.0

# Data processing
pandas==2.2.3
openpyxl==3.1.5
numpy==2.2.4

# Authentication and Security
python-decouple==3.8
python-dotenv

# Testing
pytest==8.3.5
Faker==37.1.0

# AI/ML
ollama

# HTTP and APIs
requests==2.32.3
jwt==1.3.1

# Async
aiohttp==3.11.16
aiohappyeyeballs==2.6.1
aiosignal==1.3.2
anyio==4.9.0

# Web server
gunicorn==23.0.0
uvicorn==0.34.2

# Other utilities
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2